+++ src/hcl_gpio.c
@@ -18,8 +18,11 @@
 //
 //==============================================================================
 
-#include "hcl.h"
+#include <stdint.h>
+#include <stdio.h>
+
 #include "hcl_gpio.h"
+#include "hcl.h"
 
 /*****************************************************************************
 ** Function name:       gpioInit
@@ -33,7 +36,7 @@ int gpioInit(void) { return OK; }
 ** Function name:       gpioInit
 ** Description:         Release generic GPIO interface (if any)
 ** Parameters:          None
-** Return value:        OK__attribute__((unused)) or NG
+** Return value:        OK
 *****************************************************************************/
 int gpioRelease(void) { return OK; }