+++ src/hcl_uart.h
@@ -17,8 +17,9 @@
 #define HCL_UART_H_
 
 // Common supported BAUD rates.
-// Only G365/G370/G325 can support 921600 baud.
-// 921600 baud may not be supported on all host hardware platforms.
+// Only G330PDG0/G365PDF1/G366PDG0/G370PDF1/G370PDS0/G370PDG0/G370PDT0 can
+// support 921600 baud. 921600 baud may not be supported on all host hardware
+// platforms.
 #define BAUD_921600 921600
 #define BAUD_460800 460800
 #define BAUD_230400 230400