+++ src/sensor_epsonV340.c
@@ -21,7 +21,7 @@
 ** Description:         Initialize the sensor hardware to desired settings
 **                      based on EpsonOptions
 ** Parameters:          struct EpsonOptions
-** Return value:        SUCCESS or FAIL
+** Return value:        OK or NG
 **
 *****************************************************************************/
 int sensorInitOptions(struct EpsonOptions options) {
@@ -91,7 +91,7 @@ int sensorInitOptions(struct EpsonOptions options) {
 
   registerWriteByte(CMD_WINDOW1, ADDR_COUNT_CTRL_LO, count_ctrl_lo, debug);
 
-  return TRUE;
+  return OK;
 }
 
 /*****************************************************************************